TOS-126 [Review Changes] Added CustomException
