+++ server/src/main/java/com/testsigma/config/StorageServiceFactory.java
@@ -1,5 +1,6 @@
 package com.testsigma.config;
 
+import com.testsigma.model.StorageConfig;
 import com.testsigma.service.*;
 import com.testsigma.util.HttpClient;
 import lombok.Data;
@@ -76,6 +77,23 @@ public class StorageServiceFactory {
       return true;
     return false;
 
+  }
+
+  public boolean validateCredentials(StorageConfig storageConfigReq) {
+    switch (storageConfigReq.getStorageType()) {
+      case AWS_S3:
+        if (storageConfigReq.getAwsAccessKey()==null && storageConfigReq.getAwsSecretKey()==null) return true;
+        AwsS3StorageService awsS3StorageService = new AwsS3StorageService(storageConfigReq, applicationConfig, httpClient);
+        return awsS3StorageService.validateCredentials();
+      case AZURE_BLOB:
+        if(storageConfigReq.getAzureConnectionString() == null) return true;
+        AzureBlobStorageService azureBlobStorageService = new AzureBlobStorageService(storageConfigReq, applicationConfig, httpClient);
+        return azureBlobStorageService.validateCredentials();
+      case ON_PREMISE:
+      case TESTSIGMA:
+        return true;
+    }
+    return false;
   }
 
 }
